Rename non-test files in tests directory
